Change grid-crm live-editing structure
