+++ live-editing/Routes.ts
@@ -3,8 +3,7 @@ import * as GridFinjsRouting from '../projects/app-lob/src/app/grid-finjs/grid-f
 import { GridFinjsModule } from '../projects/app-lob/src/app/grid-finjs/grid-finjs.module';
 import * as TreeGridFinjsRouting from '../projects/app-lob/src/app/treegrid-finjs/treegrid-finjs-routing.module';
 import { TreeGridFinjsModule } from '../projects/app-lob/src/app/treegrid-finjs/treegrid-finjs.module';
-
-import * as GridCRMRouting from '../projects/app-crm/src/app/grid-crm/grid-crm-routing.module';
+import * as GridCRMRouting from '../projects/app-crm/src/app/app-routing.module';
 
 import * as GridsDVRouting from '../projects/app-lob/src/app/grid/grids-routing.module';
 import { GridsDVModule } from '../projects/app-lob/src/app/grid/grids.module';
@@ -48,9 +47,7 @@ import { PaginationModule } from '../src/app/pagination/pagination.module';
 import * as PaginatorRouting from '../src/app/pagination/pagination-routing.module';
 import { PivotGridsModule } from '../src/app/pivot-grid/pivot-grids.module';
 import { PivotGridsRoutingModule } from '../src/app/pivot-grid/pivot-grids-routing.module';
-import { GridCRMComponent } from '../projects/app-crm/src/app/grid-crm/grid-crm.component';
 import { AppModule } from '../projects/app-crm/src/app/app.module';
-import { GridsCrmModule } from '../projects/app-crm/src/app/grid-crm/grid-crm.module';
 
 export const MODULE_ROUTES = [
     {
@@ -165,7 +162,7 @@ export const DV_MODULE_ROUTES = [
 
 export const CRM_MODULE_ROUTES = [
     {
-        module: GridsCrmModule,
+        module: AppModule,
         path: 'grid-crm',
         routes: GridCRMRouting.gridCrmRoutes
     }
