+++ projects/app-crm/src/app/app.module.ts
@@ -1,26 +1,37 @@
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { BrowserModule, HammerModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import {
-    IgxAutocompleteModule, IgxLayoutModule
+    IgxAutocompleteModule, IgxButtonModule, IgxIconModule, IgxInputGroupModule, IgxLayoutModule,
+    IgxRippleModule, IgxGridModule, IgxAvatarModule
 } from 'igniteui-angular';
+import { IgxPreventDocumentScrollModule } from '../../../../src/app/directives/prevent-scroll.directive';
 import { AppComponent } from './app.component';
-import { AppRoutingModule } from './app.routing.module';
-import { GridsCrmModule } from './grid-crm/grid-crm.module';
+import { GridCRMComponent } from './grid-crm/grid-crm.component';
+import { AppRoutingModule } from './app-routing.module';
 
 @NgModule({
     bootstrap: [AppComponent],
     declarations: [
-        AppComponent
+        AppComponent,
+        GridCRMComponent
     ],
     imports: [
         IgxAutocompleteModule,
+        IgxRippleModule,
+        IgxGridModule,
+        IgxIconModule,
         IgxLayoutModule,
+        IgxAvatarModule,
+        IgxInputGroupModule,
         BrowserModule,
         AppRoutingModule,
         BrowserAnimationsModule,
-        HammerModule,
-        GridsCrmModule
+        IgxButtonModule,
+        FormsModule,
+        IgxPreventDocumentScrollModule,
+        HammerModule
     ],
     providers: []
 })
