--- projects/app-crm/src/app/app.routing.module.ts
@@ -1,21 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-export const appRoutes: Routes = [
-    {
-        path: '', pathMatch: 'full', redirectTo: '/grid-crm'
-    },
-    {
-        loadChildren: () => import('./grid-crm/grid-crm.module').then(m => m.GridsCrmModule),
-        path: 'grid-crm'
-    }
-];
-@NgModule({
-    exports: [
-        RouterModule
-    ],
-    imports: [
-        RouterModule.forRoot(appRoutes)
-    ]
-})
-export class AppRoutingModule { }
